backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "April tax filing backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "browser-use>=0.1",
  "langchain-anthropic>=0.3",
  "pdfplumber>=0.11",
  "python-multipart>=0.0.12",
  "python-dotenv>=1.0",
  "pydantic-settings>=2.0",
  "aiofiles>=23.0",
  "anthropic>=0.40",
]

[project.scripts]
start = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
